requirements.txt
pandas==2.1.4
pymongo==4.6.1
python-dotenv==1.0.0
streamlit==1.26.0
streamlit_extras==0.3.6
